requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
Markdown==3.3.4
psycopg2==2.9.1
PyJWT==2.3.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
